Cleans up the features list
In the markdown file, you had each feature on it's own line. However, when this is rendered it was displaying all the features on the same line, which was making it hard to distinguish one feature from the next. I've went through and made your feature headings bold. I've also added bullets to each feature so it's easier to distinguish one from another.